requirements.txt
fastapi
uvicorn[standard]
jinja2
httpx
discord.py
python-multipart
Pillow
python-dotenv
